+++ NEWS.md
@@ -1,6 +1,7 @@
 # tidytable 0.6.7 (in development)
 
 #### Functionality improvements
+* `.data` and `.env` pronouns now work in `tidytable` functions
 * `mutate.()`: Can use `stringr::str_glue()` without specifying `.envir`
 * `replace_na.()`: Checks that `replace` arg only uses columns that exist in the data frame